Move volumereplicationgroup_controller code into its own volrep.go file
